+++ lib/main.dart
@@ -12,6 +12,7 @@ class MyApp extends StatelessWidget {
 
   // https://www.kindacode.com/article/flutter-ways-to-make-a-dark-light-mode-toggle/ - dark mode toggle
   static final ValueNotifier<ThemeMode> themeNotifier = ValueNotifier(ThemeMode.light);
+  static const appVersion = '5.0.0';
 
   // This widget is the root of your application.
   @override